requirements.txt
accelerate==0.16.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
appdirs==1.4.4
attr==0.3.1
attrs==21.2.0
autoattack @ git+https://github.com/fra31/auto-attack.git@9b264b52bb65c373373727f532865a5551ab9c02
backcall==0.2.0
beautifulsoup4==4.13.3
black==19.3b0
braceexpand==0.1.7
certifi==2021.10.8
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
click==8.0.3
cloudpickle==2.0.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.4
cycler==0.11.0
decorator==5.1.0
diffusers==0.16.0
docker-pycreds==0.4.0
einops==0.4.1
filelock==3.12.2
flake8==4.0.1
fonttools==4.42.0
fsspec==2023.6.0
ftfy==6.3.1
fvcore @ git+https://github.com/facebookresearch/fvcore.git@9d683aae73fb899dd35d6cf6720e5ef567761c57
gdown==5.1.0
geotorch==0.3.0
gitdb==4.0.10
GitPython==3.1.32
googledrivedownloader==0.4
grad-cam==1.5.5
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==2.10
imagecorruptions==1.1.2
imageio==2.10.1
importlib_metadata==8.6.1
inquirerpy==0.3.4
iopath==0.1.9
ipython==7.28.0
isort==4.3.21
jedi==0.18.0
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.4
kornia==0.6.8
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.6.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mmengine==0.10.7
mpmath==1.3.0
networkx==2.6.3
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
open_clip_torch==2.31.0
opencv-python==4.8.0.76
oss2==2.17.0
packaging==21.0
pandas==2.2.3
parameterized==0.8.1
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pfzy==0.3.4
pickleshare==0.7.5
Pillow==9.4.0
portalocker==2.3.2
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyautoattack==0.2.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.19.0
pyflakes==2.4.0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorchcv==0.0.67
pytz==2023.3.post1
PyWavelets==1.1.1
PyYAML==6.0
regex==2024.11.6
requests==2.32.3
rich==13.4.2
robustbench==1.1
safetensors==0.5.3
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.9.3
seaborn==0.12.2
sentry-sdk==1.9.0
setproctitle==1.3.2
shortuuid==1.0.11
simplejson==3.17.5
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
soupsieve==2.6
submitit==1.4.0
sympy==1.13.1
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.0.0
tifffile==2021.10.12
timm==1.0.9
tokenizers==0.13.3
toml==0.10.2
torch==2.1.0+cu121
torchdiffeq==0.2.5
torchvision==0.16.0+cu121
tqdm==4.56.2
traitlets==5.1.0
transformers==4.26.0
triton==2.1.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.25.11
wandb==0.12.17
wcwidth==0.2.5
webdataset==0.2.86
yacs==0.1.8
yapf==0.31.0
yattag==1.14.0
zipp==3.21.0
